+++ CHANGES.md
@@ -8,6 +8,7 @@
 - Broken rdl_attr_* methods
 - Bug handling optional annotated/dependent types
 - Add RDLAnnotate to rdl_disable
+- Type checking of Object class singleton methods
 
 ## [2.1.0] - 2017-06-14